refactor(racingCar): Parsing 클래스 삭제, Validation 클래스에 메소드 추가
- 메소드 쓰임이 Parsing에 적합하지 않음
- Validation 클래스에 메소드 추가
